1.修正在显示OSD时，若调用者未删除字体和文本可能会导致内存泄露的问题。 2.版本升级1.0.6.2

+++ Include/Utility.cpp
@@ -2826,11 +2826,11 @@ DWORD GetOsMajorVersion()
 	return osVer.dwMajorVersion;
 }
 
-BOOL ModifyWndStyle(HWND hWnd, int nStyleOffset,   DWORD dwRemove, DWORD dwAdd, UINT nFlags)
+BOOL ModifyWndStyle(HWND hWnd, int nStyleOffset,   long dwRemove, long dwAdd, UINT nFlags)
 {
 	assert(hWnd != NULL);
-	DWORD dwStyle = ::GetWindowLong(hWnd, nStyleOffset);
-	DWORD dwNewStyle = (dwStyle & ~dwRemove) | dwAdd;
+	long dwStyle = ::GetWindowLong(hWnd, nStyleOffset);
+	long dwNewStyle = (dwStyle & ~dwRemove) | dwAdd;
 	if (dwStyle == dwNewStyle)
 		return FALSE;
